requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
pymysql==1.0.2
werkzeug==2.3.6
cryptography>=3.3.2
reportlab==3.6.12
fpdf2==1.7.2
Pillow==9.0.1
reportlab==3.6.12